Fixed porting issue from C++ to Racket for rounding floats
In the C++, the floats are casted to integers. But in C++, this casting results in truncating numbers rather than rounding them. This is why the C++ code was multiplying by 255.99 instead of just 255. In this Racket implementation, we are using exact-round, so we need to be multiplying by 255.

This was discovered while implementing the initial code for testing the ray data type, color function, and simple camera from Chapter 3, where we were getting the wrong color.